+++ Source/ImoutoRebirth.Lilin/ImoutoRebirth.Lilin.Application/TagSlice/TagValuesSearchQuery.cs
@@ -1,0 +1,5 @@
+﻿using ImoutoRebirth.Common.Cqrs.Abstract;
+
+namespace ImoutoRebirth.Lilin.Application.TagSlice;
+
+public record TagValuesSearchQuery(Guid TagId, string? SearchPattern, int? Limit) : IQuery<IReadOnlyCollection<string>>;