+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/Services/Tags/ITagService.cs
@@ -9,7 +9,9 @@ internal interface ITagService
     Task CreateTag(Guid typeId, string name, bool hasValue, IReadOnlyCollection<string> synonyms, bool isCounter);
 
     Task<IReadOnlyCollection<Tag>> SearchTags(string name, int count, CancellationToken ct = default);
-    
+
+    Task<IReadOnlyCollection<string>> SearchTagValues(Guid tagId, string? value, int count, CancellationToken ct = default);
+
     Task<IReadOnlyCollection<Tag>> GetPopularUserTags(int count);
 
     Task<IReadOnlyCollection<Tag>> GetPopularUserCharacterTags(int count);